--- a/samples/Xamarin.Forms/CounterApp/CounterApp.iOS/CounterApp.iOS.fsproj
+++ b/samples/Xamarin.Forms/CounterApp/CounterApp.iOS/CounterApp.iOS.fsproj
@@ -26,6 +26,8 @@
         <MtouchArch>x86_64</MtouchArch>
         <MtouchLink>None</MtouchLink>
         <MtouchDebug>true</MtouchDebug>
+        <CodesignProvision>Automatic</CodesignProvision>
+        <CodesignKey>iPhone Developer</CodesignKey>
     </PropertyGroup>
     <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|iPhoneSimulator' ">
         <DebugType>none</DebugType>
@@ -50,6 +52,7 @@
         <CodesignKey>iPhone Developer</CodesignKey>
         <MtouchDebug>true</MtouchDebug>
         <CodesignEntitlements>Entitlements.plist</CodesignEntitlements>
+        <MtouchLink>SdkOnly</MtouchLink>
     </PropertyGroup>
     <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|iPhone' ">
         <DebugType>none</DebugType>
@@ -61,6 +64,7 @@
         <ConsolePause>false</ConsolePause>
         <CodesignKey>iPhone Developer</CodesignKey>
         <CodesignEntitlements>Entitlements.plist</CodesignEntitlements>
+        <MtouchLink>SdkOnly</MtouchLink>
     </PropertyGroup>
     <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Ad-Hoc|iPhone' ">
         <DebugType>none</DebugType>
@@ -71,9 +75,10 @@
         <ConsolePause>False</ConsolePause>
         <MtouchArch>ARM64</MtouchArch>
         <BuildIpa>True</BuildIpa>
-        <CodesignProvision>Automatic:AdHoc</CodesignProvision>
-        <CodesignKey>iPhone Distribution</CodesignKey>
+        <CodesignProvision>iOS Team Provisioning Profile: org.fabulous.xf.CounterApp</CodesignProvision>
+        <CodesignKey>Apple Development: Timothe Lariviere (8AK42GUU25)</CodesignKey>
         <CodesignEntitlements>Entitlements.plist</CodesignEntitlements>
+        <MtouchLink>Full</MtouchLink>
     </PropertyGroup>
     <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'AppStore|iPhone' ">
         <DebugType>none</DebugType>
@@ -86,6 +91,7 @@
         <CodesignProvision>Automatic:AppStore</CodesignProvision>
         <CodesignKey>iPhone Distribution</CodesignKey>
         <CodesignEntitlements>Entitlements.plist</CodesignEntitlements>
+        <MtouchLink>SdkOnly</MtouchLink>
     </PropertyGroup>
     <ItemGroup>
         <Reference Include="FSharp.Core, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">