--- a/samples/Xamarin.Forms/CounterApp/CounterApp.iOS/CounterApp.iOS.fsproj
+++ b/samples/Xamarin.Forms/CounterApp/CounterApp.iOS/CounterApp.iOS.fsproj
@@ -35,9 +35,10 @@
         <OutputPath>bin\iPhoneSimulator\Release</OutputPath>
         <ErrorReport>prompt</ErrorReport>
         <WarningLevel>4</WarningLevel>
-        <MtouchLink>None</MtouchLink>
+        <MtouchLink>Full</MtouchLink>
         <MtouchArch>x86_64</MtouchArch>
         <ConsolePause>false</ConsolePause>
+        <MtouchUseLlvm>true</MtouchUseLlvm>
     </PropertyGroup>
     <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|iPhone' ">
         <DebugSymbols>true</DebugSymbols>
@@ -64,7 +65,8 @@
         <ConsolePause>false</ConsolePause>
         <CodesignKey>iPhone Developer</CodesignKey>
         <CodesignEntitlements>Entitlements.plist</CodesignEntitlements>
-        <MtouchLink>SdkOnly</MtouchLink>
+        <MtouchLink>Full</MtouchLink>
+        <MtouchUseLlvm>true</MtouchUseLlvm>
     </PropertyGroup>
     <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Ad-Hoc|iPhone' ">
         <DebugType>none</DebugType>
@@ -79,6 +81,7 @@
         <CodesignKey>Apple Development: Timothe Lariviere (8AK42GUU25)</CodesignKey>
         <CodesignEntitlements>Entitlements.plist</CodesignEntitlements>
         <MtouchLink>Full</MtouchLink>
+        <MtouchUseLlvm>true</MtouchUseLlvm>
     </PropertyGroup>
     <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'AppStore|iPhone' ">
         <DebugType>none</DebugType>
@@ -88,10 +91,10 @@
         <WarningLevel>4</WarningLevel>
         <ConsolePause>False</ConsolePause>
         <MtouchArch>ARM64</MtouchArch>
-        <CodesignProvision>Automatic:AppStore</CodesignProvision>
+        <CodesignProvision></CodesignProvision>
         <CodesignKey>iPhone Distribution</CodesignKey>
         <CodesignEntitlements>Entitlements.plist</CodesignEntitlements>
-        <MtouchLink>SdkOnly</MtouchLink>
+        <MtouchLink>Full</MtouchLink>
     </PropertyGroup>
     <ItemGroup>
         <Reference Include="FSharp.Core, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
@@ -138,16 +141,16 @@
       <ImageAsset Include="Assets.xcassets\AppIcon.appiconset\ios-marketing.png">
         <Visible>false</Visible>
       </ImageAsset>
-      <ImageAsset Include="Assets.xcassets\AppIcon.appiconset\icon-60x60@2x.png">
-        <Visible>false</Visible>
-      </ImageAsset>
-      <ImageAsset Include="Assets.xcassets\AppIcon.appiconset\icon-60x60@3x.png">
+      <ImageAsset Include="Assets.xcassets\AppIcon.appiconset\icon-76x76@2x.png">
         <Visible>false</Visible>
       </ImageAsset>
       <ImageAsset Include="Assets.xcassets\AppIcon.appiconset\icon-76x76@1x.png">
         <Visible>false</Visible>
       </ImageAsset>
-      <ImageAsset Include="Assets.xcassets\AppIcon.appiconset\icon-76x76@2x.png">
+      <ImageAsset Include="Assets.xcassets\AppIcon.appiconset\icon-60x60@3x.png">
+        <Visible>false</Visible>
+      </ImageAsset>
+      <ImageAsset Include="Assets.xcassets\AppIcon.appiconset\icon-60x60@2x.png">
         <Visible>false</Visible>
       </ImageAsset>
       <ImageAsset Include="Assets.xcassets\AppIcon.appiconset\icon-83.5x83.5@2x.png">