--- a/samples/Xamarin.Forms/CounterApp/CounterApp.iOS/Info.plist
+++ b/samples/Xamarin.Forms/CounterApp/CounterApp.iOS/Info.plist
@@ -27,7 +27,7 @@
 	<key>CFBundleIdentifier</key>
 	<string>org.fabulous.xf.CounterApp</string>
 	<key>CFBundleVersion</key>
-	<string>1.0</string>
+	<string>1</string>
 	<key>UILaunchStoryboardName</key>
 	<string>LaunchScreen</string>
 	<key>CFBundleName</key>