--- a/src/Fabulous.Tests/TestUI.Attributes.fs
+++ b/src/Fabulous.Tests/TestUI.Attributes.fs
@@ -6,60 +6,60 @@ open Tests.Platform
 module Attributes =
 
     let definePressable name =
-        let key = AttributeDefinitionStore.getNextKey ()
+        Attributes.withName name
+            (fun key name ->
+              { Key = key
+                Name = name
+                Convert = id
+                ConvertValue = id
+                Compare = ScalarAttributeComparers.noCompare
+                UpdateNode =
+                    fun (newValueOpt, node) ->
 
-        let definition: ScalarAttributeDefinition<obj, obj, obj> =
-            { Key = key
-              Name = name
-              Convert = id
-              ConvertValue = id
-              Compare = ScalarAttributeComparers.noCompare
-              UpdateNode =
-                  fun (newValueOpt, node) ->
+                        let btn = node.Target :?> IButton
 
-                      let btn = node.Target :?> IButton
+                        match node.TryGetHandler<int>(key) with
+                        | ValueNone -> ()
+                        | ValueSome handlerId -> btn.RemovePressListener handlerId
 
-                      match node.TryGetHandler<int>(key) with
-                      | ValueNone -> ()
-                      | ValueSome handlerId -> btn.RemovePressListener handlerId
+                        match newValueOpt with
+                        | ValueNone -> node.SetHandler(key, ValueNone)
 
-                      match newValueOpt with
-                      | ValueNone -> node.SetHandler(key, ValueNone)
+                        | ValueSome msg ->
+                            let handler () =
+                                Attributes.dispatchMsgOnViewNode node msg
 
-                      | ValueSome msg ->
-                          let handler () =
-                              Attributes.dispatchMsgOnViewNode node msg
-
-                          let handlerId = btn.AddPressListener handler
-                          node.SetHandler<int>(key, ValueSome handlerId) }
-
-        AttributeDefinitionStore.set key definition
-        definition
+                            let handlerId = btn.AddPressListener handler
+                            node.SetHandler<int>(key, ValueSome handlerId) } : ScalarAttributeDefinition<obj, obj, obj>)
 
     // --------------- Actual Properties ---------------
     //    open Fabulous.Attributes
 
-    module Text =
-        let Record =
-            Attributes.define<bool> "Text" TestUI_ViewUpdaters.updateRecord
+    [<AbstractClass; Sealed>]
+    type Text =
+        static member Record =
+            Attributes.define<bool> "Record" TestUI_ViewUpdaters.updateRecord
 
-        let Text =
+        static member Text =
             Attributes.define<string> "Text" TestUI_ViewUpdaters.updateText
 
-
-    module TextStyle =
-        let TextColor =
+    [<AbstractClass; Sealed>]
+    type TextStyle =
+        static member TextColor =
             Attributes.define<string> "TextColor" TestUI_ViewUpdaters.updateTextColor
 
-    module Container =
-        let Children =
-            Attributes.defineWidgetCollection "Container_Children" (fun target -> (target :?> IContainer).Children)
+    [<AbstractClass; Sealed>]
+    type Container =
+        static member Children =
+            Attributes.defineWidgetCollection<TestViewElement> "Container_Children" (fun target -> (target :?> IContainer).Children :> System.Collections.Generic.IList<_>)
 
 
-    module Button =
-        let Pressed = definePressable "Button_Pressed"
+    [<AbstractClass; Sealed>]
+    type Button =
+        static member Pressed = definePressable "Button_Pressed"
 
 
-    module Automation =
-        let AutomationId =
+    [<AbstractClass; Sealed>]
+    type Automation =
+        static member AutomationId =
             Attributes.define<string> "AutomationId" TestUI_ViewUpdaters.updateAutomationId