… src/Fabulous.XamarinForms/Xamarin.Forms.Core.Attributes.fs
… src/Fabulous.XamarinForms/Xamarin.Forms.Core.fs
… src/Fabulous/AttributeDefinitions.fs
… src/Fabulous/Attributes.fs
… src/Fabulous/Memo.fs
… src/Fabulous/Reconciler.fs
… src/Fabulous/View.fs
… src/Fabulous/ViewNode.fs
… src/Fabulous/WidgetDefinitions.fs