Fixes #26 

***Short explanation (proper explanation to come later):***
Following @twop's advice in #26, I made the creation of `AttributeDefinitions` and `WidgetDefinitions` lazy by putting them in get-only static properties instead of `let` in modules.
This was how Don Syme did it in Fabulous v1.
The definitions are created only the first time we access that get-only property. Any subsequent access will only do a lookup in a cache.

Preliminary results
--
I published a version of CounterApp closer to the one in v1 with `Ad-Hoc|iPhone` configuration and `Link All` enabled.

### Estimated package sizes
| Branch | Estimated AppStore size | Improvement |
|--|--|--|
| main | 35.43 MB |- |
| improve-linking | 34.95 MB | 0.48 MB (-1%) |

### Compiled file sizes
| File | main (bytes) | improve-linking (bytes) | Improvement (bytes) |
|--|--|--|--|
CounterApp.iOS | 25 076 960 | 24 742 736 | 334 224 (-1,33%)
FSharp.Core.dll | 1 459 712 | 1 456 128 | 3 584 (-0,25%)
Fabulous.XamarinForms.dll | 358 912 | 337 920 | 20 992 (-5,85%)
FSharp.Core.aotdata.arm64 | 342 216 | 330 712 | 11 504 (-3,36%)
Fabulous.XamarinForms.aotdata.arm64 | 190 296 | 85 096 | 105 200 (-55,28%)
Fabulous.aotdata.arm64 | 172 880 | 178 432 | -5 552 (+3,21%)
Fabulous.dll | 141 824 | 141 824 | 0 (0,00%)
CounterApp.aotdata.arm64 | 71 224 | 57 608 | 13616 (-19,12%)
CounterApp.dll | 35 328 | 18 944 | 16 384 (-46,38%)
CounterApp.iOS.exe | 9 728 | 9 728 | 0 (0%)
CounterApp.iOS.aotdata.arm64 | 1 536 | 1 536 | 0 (0%)

### Benchmarks for branch main
|          Method | depth |       Mean |    Error |   StdDev |       Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|---------------- |------ |-----------:|---------:|---------:|------------:|------------:|-----------:|----------:|
| **ProcessMessages** |    **10** |   **147.8 ms** |  **1.81 ms** |  **1.60 ms** |  **40500.0000** |  **14000.0000** |  **2000.0000** |    **138 MB** |
| **ProcessMessages** |    **15** | **2,829.1 ms** | **50.57 ms** | **47.31 ms** | **505000.0000** | **168000.0000** | **69000.0000** |  **1,537 MB** |
| **CreateWidgets** |    **10** |     **535.5 μs** |     **5.08 μs** |     **4.75 μs** |   **227.5391** |  **109.3750** |         **-** |    **614 KB** |
| **CreateWidgets** |    **20** | **105,111.3 μs** | **1,492.97 μs** | **1,396.53 μs** | **22400.0000** | **4200.0000** | **1000.0000** | **75,820 KB** |

### Benchmarks for branch improve-linking
|          Method | depth |       Mean |    Error |   StdDev |       Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|---------------- |------ |-----------:|---------:|---------:|------------:|------------:|-----------:|----------:|
| **ProcessMessages** |    **10** |   **321.1 ms** |  **5.68 ms** |  **6.08 ms** |  **78000.0000** |  **30000.0000** |  **6000.0000** |    **233 MB** |
| **ProcessMessages** |    **15** | **6,643.2 ms** | **77.35 ms** | **68.57 ms** | **633000.0000** | **237000.0000** | **68000.0000** |  **2,591 MB** |
| **CreateWidgets** |    **10** |     **631.1 μs** |     **1.85 μs** |     **1.73 μs** |   **240.2344** |  **117.1875** |         **-** |    **627 KB** |
| **CreateWidgets** |    **20** | **117,790.3 μs** | **1,187.92 μs** | **1,111.18 μs** | **22800.0000** | **4800.0000** | **1200.0000** | **77,482 KB** |

This allocates way more than I expected