@twop 2022-01-14
That is an awesome data. Let me tinker about this for a little bit. I feel like this is an important one, because it can impact public API **and** architecture.