@twop 2022-01-14
I think there is a world where Attributes and Widgets carry explicit reference to a definition (plus maybe some other metadata). If that is architecturally sound and efficient that might be a way to make it tree shakable. 

@TimLariviere do you know what happens if you do this:
```fsharp
// Attributes.define is a pure function, does not modify any global state
let definitionA = Attributes.define<string> (....)
let definitionB = Attributes.define<string> (....)

// but use only "A"
let attrA = definitionA.WithValue(...)
```

If `Attributes.define` is pure will `definitionB` be linked out?