@TimLariviere 2022-01-14
>If Attributes.define is pure will definitionB be linked out?

Not 100% sure, but it should be