@TimLariviere 2022-01-14
I updated the sizes data.
Found it odd we manage to shave off 20% of the package size just by improving tree shaking in F.XF.
Turns out it's not that impressive anymore.

Guess I misconfigured the main branch and it was not using the linker.

But we still get a respectable 0.5 MB off the package size.
Which is closer to the relative size of Fabulous vs Xamarin.

Decompiler sources show that unused code is deleted 🥳 
- main: https://www.decompiler.com/jar/e281d39b3d214b4db696cefa3da8df48/Fabulous.XamarinForms.dll/Fabulous.XamarinForms.XFAttributes
- improve-linking: https://www.decompiler.com/jar/5c94c7590723442fabe34d595da9c1ee/Fabulous.XamarinForms.dll/Fabulous.XamarinForms.XFAttributes