@TimLariviere 2022-01-14
From what I understand, when we'll be able to target .NET 6.0, we'll benefit from the linker being able to remove `FSharpOptimizationData` and `FSharpSignatureData` from the F# dlls.
Which we have give us a noticeable drop in dll size.

https://github.com/dotnet/linker/pull/1040