@twop 2022-01-17
It seems that investing more time in here might be not worth it at the moment, right?

I will tinker if we can remove "Key" concept in the background, if that is indeed possible then we can tree shake unused widgets and attributes.