@TimLariviere 2022-01-17
>It seems that investing more time in here might be not worth it at the moment, right?

Yes, it's not being a major improvement. But I still want to see where I'm allocating more because I need it for #42